requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.1.8
colorama==0.4.6
comtypes==1.4.16
cryptography==46.0.5
dataclasses-json==0.6.7
distro==1.9.0
docx2txt==0.9
faiss-cpu==1.13.2
fastapi==0.135.1
filelock==3.25.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.66.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.62.3
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.83
langchain-google-genai==2.0.10
langchain-openai==0.2.14
langchain-text-splitters==0.3.11
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.3.45
llvmlite==0.46.0
MarkupSafe==3.0.3
marshmallow==3.26.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
numba==0.64.0
numpy==1.26.4
openai==1.109.1
openai-whisper==20250625
orjson==3.11.7
ormsgpack==1.12.2
packaging==25.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==4.25.8
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.13.1
pydub==0.25.1
pyparsing==3.3.2
pypiwin32==223
python-dotenv==1.2.2
python-multipart==0.0.22
pyttsx3==2.99
pywin32==311
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==65.5.0
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.5
SQLAlchemy==2.0.48
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.12.0
torch==2.10.0
tqdm==4.67.3
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
whisper==1.1.10
xxhash==3.6.0
yarl==1.23.0
zstandard==0.23.0